mysite/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
idna==3.4
numpy==1.25.2
pandas==2.0.3
pydot==1.4.2
pygraphviz==1.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wget==3.2
